% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_flood_progression.R
\name{compute_flood_stats}
\alias{compute_flood_stats}
\title{Compute Flood Progression and Crest Metrics}
\usage{
compute_flood_stats(data, DataTime, RiverWaterLevel, alert, major, minor)
}
\arguments{
\item{data}{data frame with variables date and time, river water level, alert level,
major flood level and minor flood level}

\item{RiverWaterLevel}{Variable name corresponds to the River Water Level}

\item{alert}{Alert level numeric value}

\item{major}{Major flood level numeric value}

\item{minor}{Minor flood level numeric value}

\item{DateTime}{Variable name corresponds to Date and Time (\%Y-\%m-\%d \%H:\%M)}
}
\value{
A tibble with progression and crest-related metrics
}
\description{
Computes key flood progression measures from real-time river water level data.
}
