% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_dashboard.R
\name{load_dashboard}
\alias{load_dashboard}
\title{Load dashboard}
\usage{
load_dashboard()
}
\value{
open dashboard
}
\description{
Load to explore weather- and disaster-related information associated with the Ditwah Storm
}
