% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_flood_progression.R
\name{compute_flood_stats}
\alias{compute_flood_stats}
\title{Computes number of hours in alert, minor and major levels starting from 26 Nov 2025}
\usage{
compute_flood_stats(data, DateTime, RiverWaterLevel, alert, major, minor)
}
\arguments{
\item{data}{data frame with variables date and time, river water level, alert level,
major flood level and minor flood level}

\item{DateTime}{Variable name corresponds to Date and Time (\%Y-\%m-\%d \%H:\%M)}

\item{RiverWaterLevel}{Variable name corresponds to the River Water Level}

\item{alert}{Alert level numeric value}

\item{major}{Major flood level numeric value}

\item{minor}{Minor flood level numeric value}
}
\value{
A list containing a tibble including new variable called state classifying flood level and total number of hours spent in each state
}
\description{
Classify flood levels as alert, minor and major and compute the total number of hours in each state
}
